refactor: db_api/app.py をOOPルーター分割でスリム化
- app.py (~2000行) を 253行のエントリポイントに削減
- api_common.py: 共通ヘルパー・バリデーション・レスポンスビルダーを集約
- routers/query_router.py: QueryRouter クラス（GET系エンドポイント）
- routers/governance_router.py: GovernanceRouter クラス（ガバナンス書き込み系）
- routers/ingest_router.py: IngestRouter クラス（ingest書き込み系）

後方互換性:
- db_app._parse_chart_pk, _connect_readonly, DATA_ROOT, pathlib 等を app.py で再エクスポート
- _build_waveform_preview はモンキーパッチ互換のため lambda 経由で注入

テスト: 全 db_api エンドポイントテスト通過
