+++ src/portfolio_fdc/db_api/routers/__init__.py
@@ -1,0 +1,1 @@
+"""db_api APIRouter 集約モジュール。"""
